classes: add rk-kernel-fitimage.bbclass
which is a highly modified copy of kernel-fitimage.bbclass
from OE-Core, tailored for the Rockchip targets.  This is
mainly needed to support FIT signing for the Rockchips, for
which we need to use Rockchip's version of the u-boot tools
to ensure that the extra RSA key parameters the Rk platforms
need are inserted into the device tree.

There are some other task dependency differences when signing
for Rockchip vs. what was implemented in OE-Core.

Also note that this class only implements a single FIT image
assembly function that runs after initramfs bundling, rather
than OE-Core's two functions.  The only potential downside
to this is that if you don't use an initramfs, there is no
way to install the kernel FIT image into the rootfs in
the normal way (via a package).

Signed-off-by: Matt Madison <matt@madison.systems>