wic: rkupdateimg: add padding to partitions
Using upgrade_tool and an update image to do an initial
load can sometimes result in boot failures (FIT image
checksum mismatches), while writing a wic image with the
same content works. Testing indicates that the failure happens
with partitions that are not an even multiple 1KiB in size, so
ensure that such partition files are padded with an extra sector
of zeroes before assembling the update package.

Signed-off-by: Matt Madison <matt@madison.systems>