optee-client-rockchip: add udev rules, update systemd unit
Add udev rules to set group ownership of the tee devices
to allow for access from non-root users, and trigger the
start of the tee-supplicant service when the devices are
created.

Also allow for the possibility of running the supplicant
as a non-root user in the future. Unfortunately, the
Rockchip-supplied binaries currently appear to require
root.

Signed-off-by: Matt Madison <matt@madison.systems>